+++ src/main/scala/com/gu/support/catalog/Catalog.scala
@@ -1,0 +1,5 @@
+package com.gu.support.catalog
+
+case class Catalog(
+  products: List[Product]
+)
