+++ src/main/scala/com/gu/support/catalog/PaperPrices.scala
@@ -1,0 +1,10 @@
+package com.gu.support.catalog
+
+case class PaperPrices(
+  collection: List[PricePlan],
+  delivery: List[PricePlan]
+)
+
+object PaperPrices {
+  def empty: PaperPrices = PaperPrices(List.empty, List.empty)
+}
