+++ src/main/scala/com/gu/support/catalog/PricePlan.scala
@@ -1,0 +1,7 @@
+package com.gu.support.catalog
+
+case class PricePlan(
+  id: String,
+  name: Option[String],
+  pricePerPeriod: List[Pricing]
+)
