+++ src/main/scala/com/gu/support/catalog/Product.scala
@@ -1,0 +1,7 @@
+package com.gu.support.catalog
+
+case class Product(
+  id: ProductId,
+  name: String,
+  productRatePlans: List[ProductRatePlan]
+)
