+++ src/main/scala/com/gu/support/promotions/ValidatedPromotion.scala
@@ -1,0 +1,3 @@
+package com.gu.support.promotions
+
+case class ValidatedPromotion(promoCode: PromoCode, promotion: Promotion)
